atg. elminate ctz asm. fixes #89
